Allow the eluent and the calibrant/suspect table to be specified directly as a data.frame.
These changes allow the eluent and calibrant/suspect data table to be set directly as a data.frame by the user. It is still possible to specify paths to these tables: the code checks if a path is given and then loads the table as before, otherwise the data is directly used. To accommodate these changes the relevant function parameter names are now a bit more generic.